requirements.txt
streamlit==1.28.2
pillow<11
numpy<2
opencv-python-headless<4.10
streamlit-drawable-canvas==0.9.3
pytesseract
pubchempy
streamlit-cropper